Inspections: 'probable bugs'
