+++ resources/META-INF/extensions/inspections/probablebugs/packages.xml
@@ -1,0 +1,24 @@
+<idea-plugin>
+    <extensions defaultExtensionNs="com.intellij">
+        <localInspection language="Latex" implementationClass="nl.hannahsten.texifyidea.inspections.latex.probablebugs.packages.LatexPackageCouldNotBeFound"
+                         groupName="LaTeX" displayName="Package could not be found"
+                         enabledByDefault="false"
+                         level="WARNING" />
+        <localInspection language="Latex" implementationClass="nl.hannahsten.texifyidea.inspections.latex.probablebugs.packages.LatexPackageNotInstalledInspection"
+                         groupName="LaTeX" displayName="Package is not installed"
+                         enabledByDefault="true"
+                         level="WARNING" />
+        <localInspection language="Latex" implementationClass="nl.hannahsten.texifyidea.inspections.latex.probablebugs.packages.LatexPackageNameDoesNotMatchFileNameInspection"
+                         groupName="LaTeX" displayName="Package name does not match file name"
+                         enabledByDefault="true"
+                         level="WARNING" />
+        <localInspection language="Latex" implementationClass="nl.hannahsten.texifyidea.inspections.latex.probablebugs.packages.LatexPackageSubdirectoryInspection"
+                         groupName="LaTeX" displayName="Package name does not have the correct directory"
+                         enabledByDefault="true"
+                         level="WARNING" />
+        <localInspection language="Latex" implementationClass="nl.hannahsten.texifyidea.inspections.latex.probablebugs.packages.LatexMissingImportInspection"
+                         groupName="LaTeX" displayName="Missing imports"
+                         enabledByDefault="true"
+                         level="ERROR" />
+    </extensions>
+</idea-plugin>
