+++ src/nl/hannahsten/texifyidea/inspections/latex/codestyle/spacing/LatexSentenceEndWithCapitalInspection.kt
@@ -1,4 +1,4 @@
-package nl.hannahsten.texifyidea.inspections.latex.codestyle
+package nl.hannahsten.texifyidea.inspections.latex.codestyle.spacing
 
 import com.intellij.psi.PsiElement
 import nl.hannahsten.texifyidea.inspections.TexifyRegexInspection
