+++ src/nl/hannahsten/texifyidea/inspections/latex/probablebugs/LatexAbsolutePathInspection.kt
@@ -1,4 +1,4 @@
-package nl.hannahsten.texifyidea.inspections.latex
+package nl.hannahsten.texifyidea.inspections.latex.probablebugs
 
 import com.intellij.codeInspection.InspectionManager
 import com.intellij.codeInspection.ProblemDescriptor
