+++ src/nl/hannahsten/texifyidea/inspections/latex/probablebugs/LatexNoExtensionInspection.kt
@@ -1,4 +1,4 @@
-package nl.hannahsten.texifyidea.inspections.latex
+package nl.hannahsten.texifyidea.inspections.latex.probablebugs
 
 import com.intellij.codeInspection.InspectionManager
 import com.intellij.codeInspection.LocalQuickFix
@@ -11,7 +11,6 @@ import nl.hannahsten.texifyidea.inspections.InsightGroup
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionBase
 import nl.hannahsten.texifyidea.lang.magic.MagicCommentScope
 import nl.hannahsten.texifyidea.psi.LatexCommands
-import nl.hannahsten.texifyidea.util.appendExtension
 import nl.hannahsten.texifyidea.util.files.commandsInFile
 import nl.hannahsten.texifyidea.util.files.document
 import nl.hannahsten.texifyidea.util.magic.CommandMagic
@@ -19,42 +18,42 @@ import nl.hannahsten.texifyidea.util.replaceString
 import java.util.*
 
 /**
- * See [LatexNoExtensionInspection].
+ * @author Sten Wessel
  */
-open class LatexRequiredExtensionInspection : TexifyInspectionBase() {
+open class LatexNoExtensionInspection : TexifyInspectionBase() {
 
     override val inspectionGroup = InsightGroup.LATEX
 
-    override val inspectionId = "RequiredExtension"
+    override val inspectionId = "NoExtension"
 
     override val outerSuppressionScopes = EnumSet.of(MagicCommentScope.GROUP)!!
 
-    override fun getDisplayName() = "File argument should include the extension"
+    override fun getDisplayName() = "File argument should not include the extension"
 
     override fun inspectFile(file: PsiFile, manager: InspectionManager, isOntheFly: Boolean): MutableList<ProblemDescriptor> {
         val descriptors = descriptorList()
 
         file.commandsInFile().asSequence()
-            .filter { it.name in CommandMagic.requiredExtensions }
+            .filter { it.name in CommandMagic.illegalExtensions }
             .filter { command ->
-                CommandMagic.requiredExtensions[command.name]!!.any {
+                CommandMagic.illegalExtensions[command.name]!!.any {
                     extension ->
-                    command.requiredParameters.any { !it.split(",").any { parameter -> parameter.endsWith(extension) } }
+                    command.requiredParameters.any { it.split(",").any { parameter -> parameter.endsWith(extension) } }
                 }
             }
             .forEach { command ->
                 val parameterList = command.requiredParameters.map { it.split(",") }.flatten()
-                var offset = 0
+                var offset = command.parameterList.first { it.requiredParam != null }.textOffset - command.textOffset + 1
                 for (parameter in parameterList) {
-                    if (CommandMagic.requiredExtensions[command.name]!!.any { !parameter.endsWith(it) }) {
+                    if (CommandMagic.illegalExtensions[command.name]!!.any { parameter.endsWith(it) }) {
                         descriptors.add(
                             manager.createProblemDescriptor(
                                 command,
-                                TextRange(offset, offset + parameter.length).shiftRight(command.commandToken.textLength + 1),
-                                "File argument should include the extension",
+                                TextRange(offset, offset + parameter.length),
+                                "File argument should not include the extension",
                                 ProblemHighlightType.GENERIC_ERROR_OR_WARNING,
                                 isOntheFly,
-                                AddExtensionFix
+                                RemoveExtensionFix
                             )
                         )
                     }
@@ -68,11 +67,11 @@ open class LatexRequiredExtensionInspection : TexifyInspectionBase() {
     }
 
     /**
-     * See [LatexNoExtensionInspection.RemoveExtensionFix].
+     * @author Sten Wessel
      */
-    object AddExtensionFix : LocalQuickFix {
+    object RemoveExtensionFix : LocalQuickFix {
 
-        override fun getFamilyName() = "Add file extension for parameters"
+        override fun getFamilyName() = "Remove file extension from parameters"
 
         override fun applyFix(project: Project, descriptor: ProblemDescriptor) {
             val command = descriptor.psiElement as LatexCommands
@@ -81,14 +80,14 @@ open class LatexRequiredExtensionInspection : TexifyInspectionBase() {
             val parameterList = command.requiredParameters.map { it.split(",") }.flatten()
             var offset = 0
             for (parameter in parameterList) {
-                if (CommandMagic.requiredExtensions[command.name]!!.any { !parameter.endsWith(it) }) {
+                if (CommandMagic.illegalExtensions.getOrDefault(command.name, null)?.any { parameter.endsWith(it) } == true) {
                     val range = TextRange(offset, offset + parameter.length).shiftRight(command.parameterList.first { it.requiredParam != null }.textOffset + 1)
-                    val replacement = CommandMagic.requiredExtensions[command.name]
-                        ?.find { !parameter.endsWith(it) }
-                        ?.run { parameter.appendExtension(this) } ?: break
+                    val replacement = CommandMagic.illegalExtensions[command.name]
+                        ?.find { parameter.endsWith(it) }
+                        ?.run { parameter.removeSuffix(this) } ?: break
                     document.replaceString(range, replacement)
 
-                    // Maintain offset for any added part
+                    // Maintain offset for any removed part
                     offset -= (range.length - replacement.length)
                 }
