+++ src/nl/hannahsten/texifyidea/inspections/latex/probablebugs/LatexRequiredExtensionInspection.kt
@@ -1,4 +1,4 @@
-package nl.hannahsten.texifyidea.inspections.latex
+package nl.hannahsten.texifyidea.inspections.latex.probablebugs
 
 import com.intellij.codeInspection.InspectionManager
 import com.intellij.codeInspection.LocalQuickFix
@@ -11,6 +11,7 @@ import nl.hannahsten.texifyidea.inspections.InsightGroup
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionBase
 import nl.hannahsten.texifyidea.lang.magic.MagicCommentScope
 import nl.hannahsten.texifyidea.psi.LatexCommands
+import nl.hannahsten.texifyidea.util.appendExtension
 import nl.hannahsten.texifyidea.util.files.commandsInFile
 import nl.hannahsten.texifyidea.util.files.document
 import nl.hannahsten.texifyidea.util.magic.CommandMagic
@@ -18,42 +19,42 @@ import nl.hannahsten.texifyidea.util.replaceString
 import java.util.*
 
 /**
- * @author Sten Wessel
+ * See [LatexNoExtensionInspection].
  */
-open class LatexNoExtensionInspection : TexifyInspectionBase() {
+open class LatexRequiredExtensionInspection : TexifyInspectionBase() {
 
     override val inspectionGroup = InsightGroup.LATEX
 
-    override val inspectionId = "NoExtension"
+    override val inspectionId = "RequiredExtension"
 
     override val outerSuppressionScopes = EnumSet.of(MagicCommentScope.GROUP)!!
 
-    override fun getDisplayName() = "File argument should not include the extension"
+    override fun getDisplayName() = "File argument should include the extension"
 
     override fun inspectFile(file: PsiFile, manager: InspectionManager, isOntheFly: Boolean): MutableList<ProblemDescriptor> {
         val descriptors = descriptorList()
 
         file.commandsInFile().asSequence()
-            .filter { it.name in CommandMagic.illegalExtensions }
+            .filter { it.name in CommandMagic.requiredExtensions }
             .filter { command ->
-                CommandMagic.illegalExtensions[command.name]!!.any {
+                CommandMagic.requiredExtensions[command.name]!!.any {
                     extension ->
-                    command.requiredParameters.any { it.split(",").any { parameter -> parameter.endsWith(extension) } }
+                    command.requiredParameters.any { !it.split(",").any { parameter -> parameter.endsWith(extension) } }
                 }
             }
             .forEach { command ->
                 val parameterList = command.requiredParameters.map { it.split(",") }.flatten()
-                var offset = command.parameterList.first { it.requiredParam != null }.textOffset - command.textOffset + 1
+                var offset = 0
                 for (parameter in parameterList) {
-                    if (CommandMagic.illegalExtensions[command.name]!!.any { parameter.endsWith(it) }) {
+                    if (CommandMagic.requiredExtensions[command.name]!!.any { !parameter.endsWith(it) }) {
                         descriptors.add(
                             manager.createProblemDescriptor(
                                 command,
-                                TextRange(offset, offset + parameter.length),
-                                "File argument should not include the extension",
+                                TextRange(offset, offset + parameter.length).shiftRight(command.commandToken.textLength + 1),
+                                "File argument should include the extension",
                                 ProblemHighlightType.GENERIC_ERROR_OR_WARNING,
                                 isOntheFly,
-                                RemoveExtensionFix
+                                AddExtensionFix
                             )
                         )
                     }
@@ -67,11 +68,11 @@ open class LatexNoExtensionInspection : TexifyInspectionBase() {
     }
 
     /**
-     * @author Sten Wessel
+     * See [LatexNoExtensionInspection.RemoveExtensionFix].
      */
-    object RemoveExtensionFix : LocalQuickFix {
+    object AddExtensionFix : LocalQuickFix {
 
-        override fun getFamilyName() = "Remove file extension from parameters"
+        override fun getFamilyName() = "Add file extension for parameters"
 
         override fun applyFix(project: Project, descriptor: ProblemDescriptor) {
             val command = descriptor.psiElement as LatexCommands
@@ -80,14 +81,14 @@ open class LatexNoExtensionInspection : TexifyInspectionBase() {
             val parameterList = command.requiredParameters.map { it.split(",") }.flatten()
             var offset = 0
             for (parameter in parameterList) {
-                if (CommandMagic.illegalExtensions.getOrDefault(command.name, null)?.any { parameter.endsWith(it) } == true) {
+                if (CommandMagic.requiredExtensions[command.name]!!.any { !parameter.endsWith(it) }) {
                     val range = TextRange(offset, offset + parameter.length).shiftRight(command.parameterList.first { it.requiredParam != null }.textOffset + 1)
-                    val replacement = CommandMagic.illegalExtensions[command.name]
-                        ?.find { parameter.endsWith(it) }
-                        ?.run { parameter.removeSuffix(this) } ?: break
+                    val replacement = CommandMagic.requiredExtensions[command.name]
+                        ?.find { !parameter.endsWith(it) }
+                        ?.run { parameter.appendExtension(this) } ?: break
                     document.replaceString(range, replacement)
 
-                    // Maintain offset for any removed part
+                    // Maintain offset for any added part
                     offset -= (range.length - replacement.length)
                 }
