+++ src/nl/hannahsten/texifyidea/inspections/latex/probablebugs/packages/LatexPackageNameDoesNotMatchFileNameInspection.kt
@@ -1,4 +1,4 @@
-package nl.hannahsten.texifyidea.inspections.latex
+package nl.hannahsten.texifyidea.inspections.latex.probablebugs.packages
 
 import com.intellij.codeInspection.InspectionManager
 import com.intellij.codeInspection.LocalQuickFix
