+++ test/nl/hannahsten/texifyidea/inspections/latex/LatexAbsolutePathInspectionTest.kt
@@ -2,6 +2,7 @@ package nl.hannahsten.texifyidea.inspections.latex
 
 import nl.hannahsten.texifyidea.file.LatexFileType
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionTestBase
+import nl.hannahsten.texifyidea.inspections.latex.probablebugs.LatexAbsolutePathInspection
 import org.junit.Test
 import java.nio.file.Path
 import java.nio.file.Paths
