+++ test/nl/hannahsten/texifyidea/inspections/latex/LatexEscapeAmpersandInspectionTest.kt
@@ -2,6 +2,7 @@ package nl.hannahsten.texifyidea.inspections.latex
 
 import nl.hannahsten.texifyidea.file.LatexFileType
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionTestBase
+import nl.hannahsten.texifyidea.inspections.latex.probablebugs.LatexEscapeAmpersandInspection
 import nl.hannahsten.texifyidea.util.magic.EnvironmentMagic
 
 internal class LatexEscapeAmpersandInspectionTest : TexifyInspectionTestBase(LatexEscapeAmpersandInspection()) {
