+++ test/nl/hannahsten/texifyidea/inspections/latex/LatexFileNotFoundInspectionTest.kt
@@ -2,6 +2,7 @@ package nl.hannahsten.texifyidea.inspections.latex
 
 import nl.hannahsten.texifyidea.file.LatexFileType
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionTestBase
+import nl.hannahsten.texifyidea.inspections.latex.probablebugs.LatexFileNotFoundInspection
 import org.junit.Test
 import java.io.File
 import java.nio.file.Path
