+++ test/nl/hannahsten/texifyidea/inspections/latex/LatexInclusionLoopInspectionTest.kt
@@ -1,6 +1,7 @@
 package nl.hannahsten.texifyidea.inspections.latex
 
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionTestBase
+import nl.hannahsten.texifyidea.inspections.latex.probablebugs.LatexInclusionLoopInspection
 
 class LatexInclusionLoopInspectionTest : TexifyInspectionTestBase(LatexInclusionLoopInspection()) {
