+++ test/nl/hannahsten/texifyidea/inspections/latex/LatexLabelBeforeCaptionInspectionTest.kt
@@ -1,6 +1,7 @@
 package nl.hannahsten.texifyidea.inspections.latex
 
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
+import nl.hannahsten.texifyidea.inspections.latex.probablebugs.LatexLabelBeforeCaptionInspection
 import nl.hannahsten.texifyidea.testutils.writeCommand
 import org.junit.Test
