+++ test/nl/hannahsten/texifyidea/inspections/latex/LatexMathOperatorEscapeInspectionTest.kt
@@ -1,7 +1,7 @@
 package nl.hannahsten.texifyidea.inspections.latex
 
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionTestBase
-import nl.hannahsten.texifyidea.inspections.latex.codestyle.LatexMathOperatorEscapeInspection
+import nl.hannahsten.texifyidea.inspections.latex.codestyle.spacing.LatexMathOperatorEscapeInspection
 
 class LatexMathOperatorEscapeInspectionTest : TexifyInspectionTestBase(LatexMathOperatorEscapeInspection()) {
