+++ test/nl/hannahsten/texifyidea/inspections/latex/LatexMissingDocumentclassInspectionTest.kt
@@ -1,6 +1,7 @@
 package nl.hannahsten.texifyidea.inspections.latex
 
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionTestBase
+import nl.hannahsten.texifyidea.inspections.latex.probablebugs.LatexMissingDocumentclassInspection
 
 class LatexMissingDocumentclassInspectionTest : TexifyInspectionTestBase(LatexMissingDocumentclassInspection()) {
