+++ test/nl/hannahsten/texifyidea/inspections/latex/LatexNonBreakingSpaceInspectionTest.kt
@@ -2,7 +2,7 @@ package nl.hannahsten.texifyidea.inspections.latex
 
 import nl.hannahsten.texifyidea.file.LatexFileType
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionTestBase
-import nl.hannahsten.texifyidea.inspections.latex.codestyle.LatexNonBreakingSpaceInspection
+import nl.hannahsten.texifyidea.inspections.latex.codestyle.spacing.LatexNonBreakingSpaceInspection
 import nl.hannahsten.texifyidea.testutils.writeCommand
 
 class LatexNonBreakingSpaceInspectionTest : TexifyInspectionTestBase(LatexNonBreakingSpaceInspection()) {
