+++ test/nl/hannahsten/texifyidea/inspections/latex/LatexPackageCouldNotBeFoundInspectionTest.kt
@@ -1,6 +1,7 @@
 package nl.hannahsten.texifyidea.inspections.latex
 
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
+import nl.hannahsten.texifyidea.inspections.latex.probablebugs.packages.LatexPackageCouldNotBeFound
 import org.junit.Test
 
 class LatexPackageCouldNotBeFoundInspectionTest : BasePlatformTestCase() {
