+++ test/nl/hannahsten/texifyidea/inspections/latex/LatexPackageNameDoesNotMatchFileNameInspectionTest.kt
@@ -1,6 +1,8 @@
 package nl.hannahsten.texifyidea.inspections.latex
 
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
+import nl.hannahsten.texifyidea.inspections.latex.probablebugs.packages.LatexPackageNameDoesNotMatchFileNameInspection
+import nl.hannahsten.texifyidea.inspections.latex.probablebugs.packages.LatexPackageSubdirectoryInspection
 import nl.hannahsten.texifyidea.testutils.writeCommand
 
 class LatexPackageNameDoesNotMatchFileNameInspectionTest : BasePlatformTestCase() {
