+++ test/nl/hannahsten/texifyidea/inspections/latex/LatexUnicodeInspectionTest.kt
@@ -2,6 +2,7 @@ package nl.hannahsten.texifyidea.inspections.latex
 
 import nl.hannahsten.texifyidea.file.LatexFileType
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionTestBase
+import nl.hannahsten.texifyidea.inspections.latex.probablebugs.LatexUnicodeInspection
 
 class OutsideMathLatexUnicodeInspectionTest : LatexUnicodeInspectionTest() {
